Move behavior out of layout components
