+++ src/applications/vaos/new-appointment/index.jsx
@@ -1,7 +1,15 @@
-import React from 'react';
-import { Switch, Route, useRouteMatch } from 'react-router-dom';
+import React, { useEffect } from 'react';
+import { connect } from 'react-redux';
+import {
+  Switch,
+  Route,
+  useRouteMatch,
+  useHistory,
+  useLocation,
+} from 'react-router-dom';
+import { selectIsCernerOnlyPatient } from 'platform/user/selectors';
 import newAppointmentReducer from '../reducers/newAppointment';
-import NewAppointmentLayout from '../containers/NewAppointmentLayout';
+import NewAppointmentLayout from './components/NewAppointmentLayout';
 import TypeOfCarePage from '../containers/TypeOfCarePage';
 import ContactInfoPage from '../containers/ContactInfoPage';
 import TypeOfVisitPage from '../containers/TypeOfVisitPage';
@@ -19,11 +27,61 @@ import ReviewPage from '../containers/ReviewPage';
 import ConfirmationPage from '../containers/ConfirmationPage';
 import TypeOfFacilityPage from '../containers/TypeOfFacilityPage';
 
-export function NewAppointment() {
+function onBeforeUnload(e) {
+  const expirationDate = localStorage.getItem('sessionExpiration');
+  const expirationDateSSO = localStorage.getItem('sessionExpirationSSO');
+
+  // If there's no expiration date, then the session has already expired
+  // and keeping a person on the form won't save their data
+  if (expirationDate || expirationDateSSO) {
+    e.preventDefault();
+    e.returnValue =
+      'Are you sure you wish to leave this application? All progress will be lost.';
+  }
+}
+
+function NewAppointmentSection({ isCernerOnlyPatient }) {
   const match = useRouteMatch();
+  const history = useHistory();
+  const location = useLocation();
+
+  useEffect(
+    () => {
+      if (location.pathname.endsWith('confirmation')) {
+        window.removeEventListener('beforeunload', onBeforeUnload);
+      }
+    },
+    [location],
+  );
+
+  useEffect(() => {
+    if (isCernerOnlyPatient) {
+      history.replace('/');
+    }
+
+    if (window.History) {
+      window.History.scrollRestoration = 'manual';
+    }
+
+    // We don't want people to start in the middle of the form, so redirect them when they jump
+    // in the middle. We make an exception for the confirmation page in case someone is going back
+    // after submitting.
+    if (
+      !location.pathname.endsWith('new-appointment') &&
+      !location.pathname.endsWith('confirmation')
+    ) {
+      history.replace('/new-appointment');
+    }
+
+    window.addEventListener('beforeunload', onBeforeUnload);
+
+    return () => {
+      window.removeEventListener('beforeunload', onBeforeUnload);
+    };
+  }, []);
 
   return (
-    <NewAppointmentLayout>
+    <NewAppointmentLayout isReviewPage={location.pathname.includes('review')}>
       <Switch>
         <Route path={`${match.url}/contact-info`} component={ContactInfoPage} />
         <Route
@@ -78,5 +136,13 @@ export function NewAppointment() {
     </NewAppointmentLayout>
   );
 }
+
+function mapStateToProps(state) {
+  return {
+    isCernerOnlyPatient: selectIsCernerOnlyPatient(state),
+  };
+}
+
+export const NewAppointment = connect(mapStateToProps)(NewAppointmentSection);
 
 export const reducer = newAppointmentReducer;
