+++ src/applications/vaos/tests/new-appointment/va-facility-page-multi-site.unit.spec.jsx
@@ -215,7 +215,7 @@ describe('VAOS integration: VA facility page with a multi-site user', () => {
 
     await cleanup();
 
-    screen = renderWithStoreAndRouter(<VAFacilityPage history={history} />, {
+    screen = renderWithStoreAndRouter(<VAFacilityPage />, {
       store,
     });
